Fix bugs introduced by PR
