+++ GEMstack/onboard/perception/agentformer/model/map_cnn.py
@@ -1,0 +1,32 @@
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+
+
+class MapCNN(nn.Module):
+    def __init__(self, cfg):
+        super().__init__()
+        self.convs = nn.ModuleList()
+        map_channels = cfg.get('map_channels', 3)
+        patch_size = cfg.get('patch_size', [100, 100])
+        hdim = cfg.get('hdim', [32, 32])
+        kernels = cfg.get('kernels', [3, 3])
+        strides = cfg.get('strides', [3, 3])
+        self.out_dim = out_dim = cfg.get('out_dim', 32)
+        self.input_size = input_size = (map_channels, patch_size[0], patch_size[1])
+        x_dummy = torch.randn(input_size).unsqueeze(0)
+
+        for i, _ in enumerate(hdim):
+            self.convs.append(nn.Conv2d(map_channels if i == 0 else hdim[i-1],
+                                        hdim[i], kernels[i],
+                                        stride=strides[i]))
+            x_dummy = self.convs[i](x_dummy)
+
+        self.fc = nn.Linear(x_dummy.numel(), out_dim)
+
+    def forward(self, x):
+        for conv in self.convs:
+            x = F.leaky_relu(conv(x), 0.2)
+        x = torch.flatten(x, start_dim=1)
+        x = self.fc(x)
+        return x
