+++ GEMstack/onboard/perception/agentformer/model/model_lib.py
@@ -1,0 +1,8 @@
+from model.agentformer import AgentFormer
+from model.dlow import DLow
+
+
+model_dict = {
+    'agentformer': AgentFormer,
+    'dlow': DLow
+}
