--- GEMstack/onboard/prediction/agentformer/model/agentformer_loss.py
@@ -1,43 +1,0 @@
-
-
-def compute_motion_mse(data, cfg):
-    diff = data['fut_motion_orig'] - data['train_dec_motion']
-    if cfg.get('mask', True):
-        mask = data['fut_mask']
-        diff *= mask.unsqueeze(2)
-    loss_unweighted = diff.pow(2).sum() 
-    if cfg.get('normalize', True):
-        loss_unweighted /= diff.shape[0]
-    loss = loss_unweighted * cfg['weight']
-    return loss, loss_unweighted
-
-
-def compute_z_kld(data, cfg):
-    loss_unweighted = data['q_z_dist'].kl(data['p_z_dist']).sum()
-    if cfg.get('normalize', True):
-        loss_unweighted /= data['batch_size']
-    loss_unweighted = loss_unweighted.clamp_min_(cfg.min_clip)
-    loss = loss_unweighted * cfg['weight']
-    return loss, loss_unweighted
-
-
-def compute_sample_loss(data, cfg):
-    diff = data['infer_dec_motion'] - data['fut_motion_orig'].unsqueeze(1)
-    if cfg.get('mask', True):
-        mask = data['fut_mask'].unsqueeze(1).unsqueeze(-1)
-        diff *= mask
-    dist = diff.pow(2).sum(dim=-1).sum(dim=-1)
-    loss_unweighted = dist.min(dim=1)[0]
-    if cfg.get('normalize', True):
-        loss_unweighted = loss_unweighted.mean()
-    else:
-        loss_unweighted = loss_unweighted.sum()
-    loss = loss_unweighted * cfg['weight']
-    return loss, loss_unweighted
-
-
-loss_func = {
-    'mse': compute_motion_mse,
-    'kld': compute_z_kld,
-    'sample': compute_sample_loss
-}
