--- GEMstack/onboard/prediction/agentformer/model/model_lib.py
@@ -1,8 +1,0 @@
-from model.agentformer import AgentFormer
-from model.dlow import DLow
-
-
-model_dict = {
-    'agentformer': AgentFormer,
-    'dlow': DLow
-}
