--- GEMstack/onboard/prediction/agentformer/model_cfg/inference.yml
@@ -1,57 +1,0 @@
-# ------------------- General Options -------------------------
-
-description                  : AgentFormer (DLow)
-results_root_dir             : results
-seed                         : 1
-dataset                      : eth
-data_root_ethucy             : datasets/eth_ucy
-data_root_nuscenes_pred      : datasets/nuscenes_pred
-model_path                   : model_weights/eth_inference.p
-
-# ------------------- Feature Extractor -------------------------
-
-past_frames                  : 8
-future_frames                : 12
-min_past_frames              : 8
-min_future_frames            : 12
-
-traj_scale                   : 2
-motion_dim                   : 2
-forecast_dim                 : 2
-
-# ------------------- Model -------------------------
-
-model_id: dlow
-pred_cfg: model_cfg/inference_pre.yml
-pred_epoch: 30
-qnet_mlp: [512, 256]
-share_eps: true
-train_w_mean: true
-best_samples: [0,1,17,18,19]
-
-# ------------------- VAE-------------------------
-
-nz                           : 32
-sample_k                     : 20
-
-# ------------------- Training Parameters -------------------------
-
-lr                           : 1.e-4
-loss_cfg:
-  kld:
-    weight: 0.1
-    min_clip: 10
-  diverse:
-    weight: 20
-    d_scale: 10
-  recon:
-    weight: 5.0
-
-num_epochs                   : 50
-lr_fix_epochs                : 10
-lr_scheduler: 'step'
-decay_step: 5
-decay_gamma: 0.5
-print_freq                   : 20
-model_save_freq              : 5
-
