--- GEMstack/onboard/prediction/agentformer/model_cfg/inference_pre.yml
@@ -1,75 +1,0 @@
-# ------------------- General Options -------------------------
-
-description                  : AgentFormer (Before DLow)
-results_root_dir             : results
-seed                         : 1
-dataset                      : eth
-data_root_ethucy             : datasets/eth_ucy
-data_root_nuscenes_pred      : datasets/nuscenes_pred
-model_path                   : model_weights/eth_inference_pre.p
-
-# ------------------- Feature Extractor -------------------------
-
-past_frames                  : 8
-future_frames                : 12
-min_past_frames          -------------------------------------------------------------------------------------------------------------------    : 8
-min_future_frames            : 12
-
-traj_scale                   : 2
-motion_dim                   : 2
-forecast_dim                 : 2
-
-# ------------------- Model -------------------------
-
-model_id: agentformer
-tf_version: v2
-tf_model_dim: 256
-tf_ff_dim: 512
-tf_nhead: 8
-tf_dropout: 0.1
-input_type: ['scene_norm', 'vel']
-pred_type: 'scene_norm'
-sn_out_type: 'norm'
-max_train_agent: 32
-pos_concat: true
-rand_rot_scene: true
-scene_orig_all_past: true
-
-context_encoder:
-  nlayer: 2
-
-future_decoder:
-  nlayer: 2
-  out_mlp_dim: [512, 256]
-
-future_encoder:
-  nlayer: 2
-  out_mlp_dim: [512, 256]
-
-# ------------------- VAE-------------------------
-
-nz                           : 32
-sample_k                     : 20
-learn_prior                  : true
-
-# ------------------- Training Parameters -------------------------
-
-lr                           : 1.e-4
-loss_cfg:
-  mse:
-    weight: 1.0
-  kld:
-    weight: 1.0
-    min_clip: 2.0
-  sample:
-    weight: 1.0
-    k: 20
-
-num_epochs                   : 100
-lr_fix_epochs                : 10
-lr_scheduler: 'step'
-decay_step: 10
-decay_gamma: 0.5
-print_freq                   : 20
-model_save_freq              : 10
-
