--- GEMstack/onboard/prediction/agentformer/requirements.txt
@@ -1,6 +1,0 @@
-# tensorboard
-opencv-python
-# glob2
-# six
-pyyaml
-easydict
