--- GEMstack/onboard/prediction/agentformer/utils/torch.py
@@ -1,204 +1,0 @@
-import torch
-import numpy as np
-from torch.optim import lr_scheduler
-from torch import nn
-
-tensor = torch.tensor
-DoubleTensor = torch.DoubleTensor
-FloatTensor = torch.FloatTensor
-LongTensor = torch.LongTensor
-ByteTensor = torch.ByteTensor
-ones = torch.ones
-zeros = torch.zeros
-
-
-def rotation_2d_torch(x, theta, origin=None):
-    if origin is None:
-        origin = torch.zeros(2).to(x.device).to(x.dtype)
-    norm_x = x - origin
-    norm_rot_x = torch.zeros_like(x)
-    norm_rot_x[..., 0] = norm_x[..., 0] * torch.cos(theta) - norm_x[..., 1] * torch.sin(theta)
-    norm_rot_x[..., 1] = norm_x[..., 0] * torch.sin(theta) + norm_x[..., 1] * torch.cos(theta)
-    rot_x = norm_rot_x + origin
-    return rot_x, norm_rot_x
-
-
-class to_cpu:
-
-    def __init__(self, *models):
-        self.models = list(filter(lambda x: x is not None, models))
-        self.prev_devices = [x.device if hasattr(x, 'device') else next(x.parameters()).device for x in self.models]
-        for x in self.models:
-            x.to(torch.device('cpu'))
-
-    def __enter__(self):
-        pass
-
-    def __exit__(self, *args):
-        for x, device in zip(self.models, self.prev_devices):
-            x.to(device)
-        return False
-
-
-class to_device:
-
-    def __init__(self, device, *models):
-        self.models = list(filter(lambda x: x is not None, models))
-        self.prev_devices = [x.device if hasattr(x, 'device') else next(x.parameters()).device for x in self.models]
-        for x in self.models:
-            x.to(device)
-
-    def __enter__(self):
-        pass
-
-    def __exit__(self, *args):
-        for x, device in zip(self.models, self.prev_devices):
-            x.to(device)
-        return False
-
-
-class to_test:
-
-    def __init__(self, *models):
-        self.models = list(filter(lambda x: x is not None, models))
-        self.prev_modes = [x.training for x in self.models]
-        for x in self.models:
-            x.train(False)
-
-    def __enter__(self):
-        pass
-
-    def __exit__(self, *args):
-        for x, mode in zip(self.models, self.prev_modes):
-            x.train(mode)
-        return False
-
-
-class to_train:
-
-    def __init__(self, *models):
-        self.models = list(filter(lambda x: x is not None, models))
-        self.prev_modes = [x.training for x in self.models]
-        for x in self.models:
-            x.train(True)
-
-    def __enter__(self):
-        pass
-
-    def __exit__(self, *args):
-        for x, mode in zip(self.models, self.prev_modes):
-            x.train(mode)
-        return False
-
-
-def batch_to(dst, *args):
-    return [x.to(dst) if x is not None else None for x in args]
-
-
-def get_flat_params_from(models):
-    if not hasattr(models, '__iter__'):
-        models = (models, )
-    params = []
-    for model in models:
-        for param in model.parameters():
-            params.append(param.data.view(-1))
-
-    flat_params = torch.cat(params)
-    return flat_params
-
-
-def set_flat_params_to(model, flat_params):
-    prev_ind = 0
-    for param in model.parameters():
-        flat_size = int(np.prod(list(param.size())))
-        param.data.copy_(
-            flat_params[prev_ind:prev_ind + flat_size].view(param.size()))
-        prev_ind += flat_size
-
-
-def get_flat_grad_from(inputs, grad_grad=False):
-    grads = []
-    for param in inputs:
-        if grad_grad:
-            grads.append(param.grad.grad.view(-1))
-        else:
-            if param.grad is None:
-                grads.append(zeros(param.view(-1).shape))
-            else:
-                grads.append(param.grad.view(-1))
-
-    flat_grad = torch.cat(grads)
-    return flat_grad
-
-
-def compute_flat_grad(output, inputs, filter_input_ids=set(), retain_graph=False, create_graph=False):
-    if create_graph:
-        retain_graph = True
-
-    inputs = list(inputs)
-    params = []
-    for i, param in enumerate(inputs):
-        if i not in filter_input_ids:
-            params.append(param)
-
-    grads = torch.autograd.grad(output, params, retain_graph=retain_graph, create_graph=create_graph)
-
-    j = 0
-    out_grads = []
-    for i, param in enumerate(inputs):
-        if i in filter_input_ids:
-            out_grads.append(zeros(param.view(-1).shape))
-        else:
-            out_grads.append(grads[j].view(-1))
-            j += 1
-    grads = torch.cat(out_grads)
-
-    for param in params:
-        param.grad = None
-    return grads
-
-def set_optimizer_lr(optimizer, lr):
-    for param_group in optimizer.param_groups:
-        param_group['lr'] = lr
-
-def filter_state_dict(state_dict, filter_keys):
-    for key in list(state_dict.keys()):
-        for f_key in filter_keys:
-            if f_key in key:
-                del state_dict[key]
-                break
-
-def get_scheduler(optimizer, policy, nepoch_fix=None, nepoch=None, decay_step=None, decay_gamma=0.1):
-    if policy == 'lambda':
-        def lambda_rule(epoch):
-            lr_l = 1.0 - max(0, epoch - nepoch_fix) / float(nepoch - nepoch_fix + 1)
-            return lr_l
-        scheduler = lr_scheduler.LambdaLR(optimizer, lr_lambda=lambda_rule)
-    elif policy == 'step':
-        scheduler = lr_scheduler.StepLR(
-            optimizer, step_size=decay_step, gamma=decay_gamma)
-    elif policy == 'plateau':
-        scheduler = lr_scheduler.ReduceLROnPlateau(
-            optimizer, mode='min', factor=0.2, threshold=0.01, patience=5)
-    else:
-        return NotImplementedError('learning rate policy [%s] is not implemented', policy)
-    return scheduler
-
-
-class ExpParamAnnealer(nn.Module):
-
-    def __init__(self, start, finish, rate, cur_epoch=0):
-        super().__init__()
-        self.register_buffer('start', torch.tensor(start))
-        self.register_buffer('finish', torch.tensor(finish))
-        self.register_buffer('rate', torch.tensor(rate))
-        self.register_buffer('cur_epoch', torch.tensor(cur_epoch))
-
-    def step(self):
-        self.cur_epoch += 1
-
-    def set_epoch(self, epoch):
-        self.cur_epoch.fill_(epoch)
-
-    def val(self):
-        return self.finish - (self.finish - self.start) * (self.rate ** self.cur_epoch)
